Prepare to support mutiple output options
The big part of this is preparing for builder instance to be
optional by splitting it into the docker-buildx-create rule.

No longer "docker buildx use" the builder instance and instead
explicitly pass it in with --builder when we build.

Test for and create (if needed) the context and builder
instance separately. Also make sure we cleanup both.