Add 'image' goal to store the image in the local repository
To do this, we DO NOT want to use the builder instance. The
docker-buildx-create rule is now optional and is only called
if DOCKERX_INSTANCE is set. We need to make sure we are using
the default 'docker' driver, NOT 'docker-container' so the
image ends up in the right place.