Make sure we can clean up 'make image' images
We add a timestamped "buildx" tag, so we can cleanup images later.
Without this, if you run 'make image' two times in a row, all of the
tags on the first image will be removed when the second image is
created and then moved to the second image, leaving an untagged
image in the local repository. This will at least give us an easy
way to cleanup those orphaned untagged images in 'uninstall-all'.